Backend-app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pydantic==2.5.3
passlib==1.7.4
python-jose[cryptography]==3.3.0
